requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.3
